fix: Email 타입 제거에 따른 코드 변경
