--- app-api/src/main/java/com/aireview/review/login/LoginResponse.java
@@ -1,16 +1,0 @@
-package com.aireview.review.login;
-
-import lombok.Getter;
-
-@Getter
-public class LoginResponse {
-    private final String accessToken;
-    private final String refreshToken;
-    private final Long userId;
-
-    public LoginResponse(String accessToken, String refreshToken, Long userId) {
-        this.accessToken = accessToken;
-        this.refreshToken = refreshToken;
-        this.userId = userId;
-    }
-}
