+++ app-api/src/main/java/com/aireview/review/login/LoginSuccessHandler.java
@@ -2,7 +2,8 @@ package com.aireview.review.login;
 
 import com.aireview.review.authentication.jwt.Jwt;
 import com.aireview.review.authentication.jwt.JwtService;
-import com.aireview.review.domain.user.User;
+import com.aireview.review.domains.user.domain.User;
+import com.aireview.review.model.LoginResponse;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
@@ -27,7 +28,7 @@ public abstract class LoginSuccessHandler implements AuthenticationSuccessHandle
         Long userId = user.getId();
         Jwt jwt = jwtService.createJwt(
                 userId,
-                user.getEmail().getAddress(),
+                user.getEmail(),
                 authentication.getAuthorities().stream().map(GrantedAuthority::getAuthority).toArray(String[]::new));
 
         LoginResponse loginResponse = new LoginResponse(jwt.getAccessToken(), jwt.getRefreshToken(), userId);
@@ -36,17 +37,12 @@ public abstract class LoginSuccessHandler implements AuthenticationSuccessHandle
 
     protected abstract User processAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication);
 
-    private void writeLoginSuccessResponse(HttpServletResponse response, LoginResponse loginResponse) {
-        try {
-            response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-            response.getWriter().write(
-                    objectMapper.writeValueAsString(loginResponse)
-            );
-            response.getWriter().flush();
-            response.getWriter().close();
-        } catch (IOException exception) {
-            // TODO: 5/9/24  oauth user 정보가 없으면? IllegalArgumentException 어떻게 처리?
-            // TODO: 5/9/24  user 정보를 가져오다가 예외가 발생하면 어떻게 처리할 것인가? DataAccessException 어떻게 처리?
-        }
+    private void writeLoginSuccessResponse(HttpServletResponse response, LoginResponse loginResponse) throws IOException {
+        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
+        response.getWriter().write(
+                objectMapper.writeValueAsString(loginResponse)
+        );
+        response.getWriter().flush();
+        response.getWriter().close();
     }
 }
